yask: Move stuff from __init__ to kernel.py
